fix: split overlay Transfer tests to pass app test size lint (#4625)
Quality failed repo.app_test_file_size because the overlay Transfer suite and fleet-marker flow tests exceeded 380 physical lines.
