+++ app/test/naval_fleet_marker_flow_test.dart
@@ -1,19 +1,15 @@
-// Fleet-marker routing + Sail row tests (Refs #4343).
+// Fleet-marker routing tests (Refs #4343, #4625).
 // SPEC: SPEC/ui/map-widget.md, naval-mission-menu-dialog.md, move-fleet-dialog.md.
 
 import 'package:colonizethis_app/features/game/widgets/unit_orders/in_port_fleet_marker_actions_dialog.dart';
 import 'package:colonizethis_app/features/game/widgets/unit_orders/move_fleet_dialog.dart';
 import 'package:colonizethis_app/features/game/widgets/unit_orders/naval_fleet_marker_flow.dart';
-import 'package:colonizethis_app/features/game/widgets/unit_orders/naval_mission_flow.dart';
 import 'package:colonizethis_app/features/game/widgets/unit_orders/naval_mission_menu_dialog.dart';
 import 'package:colonizethis_app/features/game/widgets/unit_orders/split_fleet_dialog.dart';
 import 'package:colonizethis_app/features/game/widgets/unit_orders/transfer_to_home_fleet_dialog.dart';
-import 'package:colonizethis_app_fixtures/config/ct_e2e.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_orders/colonizethis_orders.dart'
-    show navalMissionAvailabilityForFleet;
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -49,22 +45,6 @@ void main() {
   );
 
   Game atSeaPeerGame() => buildNavalPanelNamedSeaZoneGame(humanId: humanId);
-
-  MapTopology adjacentSeasTopology() => const MapTopology(
-    nodes: [
-      TopologyNode(
-        id: 'zone_alpha',
-        regionId: 'oldWorld',
-        type: TopologyNodeType.seaZone,
-      ),
-      TopologyNode(
-        id: 'zone_beta',
-        regionId: 'oldWorld',
-        type: TopologyNodeType.seaZone,
-      ),
-    ],
-    edges: [TopologyEdge(id1: 'zone_alpha', id2: 'zone_beta')],
-  );
 
   Future<void> pumpOpenButton(
     WidgetTester tester, {
@@ -339,104 +319,6 @@ void main() {
       expect(find.byType(NavalMissionMenuDialog), findsOneWidget);
       expect(find.text('Sail / Move'), findsOneWidget);
       expect(find.text('Patrol'), findsOneWidget);
-    });
-  });
-
-  group('DLG31001 Sail / Move', () {
-    testWidgets('menu always lists Sail / Move', (tester) async {
-      final game = atSeaPeerGame();
-      final fleet = game.worldState.fleets.single;
-      final availability = navalMissionAvailabilityForFleet(
-        game: game,
-        topology: const MapTopology(),
-        playerId: humanId,
-        fleet: fleet,
-        currentOrders: const Orders(),
-      );
-
-      await tester.pumpWidget(
-        buildAppShell(
-          child: NavalMissionMenuDialog(
-            game: game,
-            fleet: fleet,
-            availability: availability,
-          ),
-        ),
-      );
-      await tester.pumpAndSettle();
-      expect(find.text('Sail / Move'), findsOneWidget);
-      expect(
-        find.text('Move this fleet to an adjacent sea zone or owned port.'),
-        findsOneWidget,
-      );
-    });
-
-    testWidgets('confirming Sail move clears pending mission (XOR)', (
-      tester,
-    ) async {
-      await bindFixedTestSurface(tester, const Size(800, 1200));
-      final game = atSeaPeerGame();
-      final bus = AppEventBus();
-      var orders = Orders(
-        navalMissionOrdersByPlayerId: {
-          humanId: [NavalMissionOrder(fleetId: 'sea_named', mission: 'patrol')],
-        },
-      );
-      bus.on<NavalMoveFleetRequestedEvent>().listen((e) {
-        orders = applyNavalMoveOrderForPlayer(
-          orders,
-          e.humanPlayerId,
-          e.moveOrder,
-        );
-      });
-
-      final fleet = game.worldState.fleets.single;
-      await pumpOpenButton(
-        tester,
-        onPressed: (context) => showNavalMissionFlow(
-          context: context,
-          game: game,
-          topology: adjacentSeasTopology(),
-          humanPlayerId: humanId,
-          draftOrders: orders,
-          bus: bus,
-          fleetIds: [fleet.id],
-        ),
-      );
-      await tester.tap(find.text('open-flow'));
-      await tester.pumpAndSettle();
-
-      final sailFinder = find.text('Sail / Move');
-      await tester.ensureVisible(sailFinder);
-      await tester.tap(sailFinder);
-      await tester.pumpAndSettle();
-      expect(find.byType(MoveFleetDialog), findsOneWidget);
-
-      // Prefer e2e keys when CT_E2E is on; otherwise tap the destination label.
-      final seaRow = find.byKey(
-        kCtE2EMoveFleetDestinationSeaZoneRowKey('zone_beta'),
-      );
-      if (seaRow.evaluate().isNotEmpty) {
-        await tester.ensureVisible(seaRow);
-        await tester.tap(seaRow);
-      } else {
-        final label = find.text('zone_beta');
-        await tester.ensureVisible(label);
-        await tester.tap(label);
-      }
-      await tester.pumpAndSettle();
-      await tester.tap(find.text('Confirm'));
-      await tester.pumpAndSettle();
-
-      expect(orders.navalMissionOrdersByPlayerId[humanId] ?? const [], isEmpty);
-      expect(
-        orders.navalMoveOrdersByPlayerId[humanId]?.single.fleetId,
-        'sea_named',
-      );
-      expect(
-        orders.navalMoveOrdersByPlayerId[humanId]?.single.destinationSeaZoneId,
-        'zone_beta',
-      );
     });
   });
 }
